chore(compiler): move type to eliminate circular imports
