+++ packages/@lwc/compiler/src/index.ts
@@ -6,7 +6,7 @@
  */
 export { transform, transformSync } from './transformers/transformer';
 
-export { TransformResult } from './transformers/transformer';
+export { TransformResult } from './transformers/shared';
 export {
     NormalizedTransformOptions,
     TransformOptions,
